Raise correct exception when element is not found
Previously, if the element was not found in the page when it was
expected to be visible, the sentence would incorrectly raise the
exception that the element was expected to not be visisble but it was.

This fix correctly observes the `not_` parameter when the element is not
found on the page.